requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
markdown2==2.3.8
MarkupSafe==1.1.1
mongoengine==0.19.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
